scrum-board-visi/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
contourpy==1.3.0
cycler==0.12.1
distro==1.9.0
Django==5.1.1
django-cors-headers==4.5.0
djangorestframework==3.15.2
fonttools==4.54.1
groq==0.12.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.7
langchain-core==0.3.19
langchain-groq==0.2.1
langsmith==0.1.143
matplotlib==3.9.2
numpy==2.1.1
orjson==3.10.11
packaging==24.1
pandas==2.2.3
pillow==10.4.0
psycopg2==2.9.10
psycopg2-binary==2.9.6
pydantic==2.9.2
pydantic_core==2.23.4
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
tenacity==9.0.0
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
xgboost==2.1.1
